[BTAT-1960] Refactored existing component tests to move Calc Breakdown to correct Wireock Stub Helper
